+++ agentlib/core/errors.py
@@ -30,7 +30,7 @@ class OptionalDependencyError(Exception):
         message = (
             f"{used_object} is an optional dependency which you did not "
             f"install yet. Install the missing dependency "
-            f"either using `pip install {dependency_install}`"
+            f"using `pip install {dependency_install}`"
         )
         if dependency_name is not None:
             message += (